+++ src/stories/Combobox/Combobox.settings.ts
@@ -392,4 +392,12 @@ export const argTypes = {
 			},
 		},
 	},
+	'update:search': {
+		table: {
+			category: 'Events',
+			type: {
+				summary: 'string',
+			},
+		},
+	},
 }